==> 1- Data Preprocessing.ipynb <==
# Data Preprocessing
Purpose: Convert text file into readable csv file

Dataset 1
- Train: 20631 rows, 26 columns
- Test: 100 rows, 26 columns

Dataset 2
- Train: 53760 rows, 26 columns
- Test: 260 rows, 26 columns

Dataset 3
- Train: 24721 rows, 26 columns
- Test: 101 rows, 26 columns

Dataset 4
- Train: 61250 rows, 26 columns
- Test: 248 rows, 26 columns

- Regression models: How many more cycles an engine will function before it fails?

## Import

# Import Required Library
import numpy as np
import statistics
import pandas as pd
import os

## Get Working Directory

# Get Current Direcotry
os.getcwd()

# Dataset Column Name
names = ['id', 'cycle', 'setting1', 'setting2', 'setting3', 's1', 's2', 's3','s4', 's5', 's6', 's7', 's8',
         's9', 's10', 's11', 's12', 's13', 's14', 's15', 's16', 's17', 's18', 's19', 's20', 's21']

# Clean Training Dataset

# Read training data
train_data = pd.read_csv('Dataset/train_FD001.txt', sep=" ", header=None)
train_data.head()

# Drop extra space column
train_data.drop([26,27], axis=1, inplace=True)

# Assign the column names
train_data.columns = names

# First n row
train_data.head()

# Show data types
train_data.dtypes

# Check for null value
train_data.isnull().sum()

# Describe the data
train_data.describe()

# Clean Test Dataset

# Read test data
test_data = pd.read_csv('Dataset/test_FD001.txt', sep=" ", header=None)
test_data.head()

# Drop extra space column
test_data.drop([26,27], axis=1, inplace=True)

# Assign the column names
test_data.columns = names

# First n rows
test_data.head()

# Show data types
test_data.dtypes

# Check null value
test_data.isnull().sum()

# Calculate Stastical Data
test_data.describe()

# Clean Truth Dataset

# Read the truth data
truth_data = pd.read_csv('Dataset/RUL_FD001.txt', sep=" ", header=None)
truth_data.head()

# drop extra space column
truth_data.drop([1], axis=1, inplace=True)

# Assign the column names
truth_data.columns=['RUL']

truth_data.head()

# Show data types
truth_data.dtypes

# Check for null value
truth_data.isnull().sum()

truth_data.describe()

print("This is the size of the train dataset: {} rows and {} features".format(train_data.shape[0], 
                                                                                 train_data.shape[1]))
print("This is the size of the test dataset: {} rows and {} features".format(test_data.shape[0],
                                                                                test_data.shape[1]))
print("This is the size of the truth dataset: {} rows and {} features".format(truth_data.shape[0],
                                                                                 truth_data.shape[1]))

n_turb = train_data["id"].unique().max()
n_train, n_features = train_data.shape
print("There is {} turbines in the dataset".format(n_turb))

# Feature Extraction


# Create a dataframe to hold the last cycle for each turbine engine in the dataset
df_max_cycle = pd.DataFrame(train_data.groupby('id')['cycle'].max())
df_max_cycle.reset_index(level=0, inplace=True)
df_max_cycle.columns = ['id', 'last_cycle']
    
# Add Remaining Useful Life (RUL) as a new column for regression label
train_data = pd.merge(train_data, df_max_cycle, on='id')
train_data['RUL'] = train_data['last_cycle'] - train_data['cycle']
train_data.drop(['last_cycle'], axis=1, inplace=True)

train_data.dtypes

# Save to CSV
train_data.to_csv('Dataset/train.csv', index=False)
print("Train Data saved as Dataset/train.csv")

# For Test data

# Create a dataframe to hold the last cycle for each turbine engine in the dataset
df_tst_last_cycle = pd.DataFrame(test_data.groupby('id')['cycle'].max())
df_tst_last_cycle.reset_index(level=0, inplace=True)
df_tst_last_cycle.columns = ['id', 'last_cycle']

# Add Remaining Useful Life (RUL) as a new column for regression label   
test_data = pd.merge(test_data, df_tst_last_cycle, on='id')
test_data = test_data[test_data['cycle'] == test_data['last_cycle']]
test_data.drop(['last_cycle'], axis=1, inplace=True)
    
test_data.reset_index(drop=True, inplace=True)
    
test_data = pd.concat([test_data, truth_data], axis=1)

# return df_test_in

test_data.dtypes

# Save to CSV
test_data.to_csv('Dataset/test.csv', index=False)
print("Test Data saved as Dataset/test.csv")
