==> 2 - Regression.ipynb <==
# Import Required Library
import pandas as pd
import numpy as np


import matplotlib.pyplot as plt
plt.style.use('ggplot')
%matplotlib inline  

from sklearn import linear_model
from sklearn.ensemble import RandomForestRegressor
from sklearn.preprocessing import PolynomialFeatures

from sklearn import model_selection #import cross_val_score, StratifiedKFold
from sklearn.tree import DecisionTreeRegressor, DecisionTreeClassifier, export_graphviz
from sklearn import metrics  # mean_squared_error, mean_absolute_error, median_absolute_error, explained_variance_score, r2_score
from sklearn.feature_selection import SelectFromModel, RFECV
from sklearn.externals import joblib

import pickle

# load training data 
df_train = pd.read_csv('Dataset/train.csv')
df_train.head()

# load test data
df_test = pd.read_csv('Dataset/test.csv')
df_test.head()

# Declare Features
features_orig = ['setting1','setting2','setting3','s1','s2','s3','s4','s5','s6','s7','s8','s9','s10','s11','s12','s13','s14','s15','s16','s17','s18','s19','s20','s21']

features = features_orig


X_train = df_train[features]
y_train = df_train['RUL']

X_test = df_test[features]
y_test = df_test['RUL']

# Linear Regression

linreg = linear_model.LinearRegression()
linreg.fit(X_train, y_train)

y_test_predict = linreg.predict(X_test)
y_train_predict = linreg.predict(X_train)

print('R^2 training: %.3f, R^2 test: %.3f' % (
      (metrics.r2_score(y_train, y_train_predict)), 
      (metrics.r2_score(y_test, y_test_predict))))

print("Root Mean Squared Error : " + (metrics.mean_squared_error(y_test, y_test_predict)**0.5).astype(str) )

# LASSO Regression

lasso = linear_model.Lasso(alpha=0.01)
lasso.fit(X_train, y_train)

y_test_predict = lasso.predict(X_test)
y_train_predict = lasso.predict(X_train)

print('R^2 training: %.3f, R^2 test: %.3f' % (
      (metrics.r2_score(y_train, y_train_predict)), 
      (metrics.r2_score(y_test, y_test_predict))))

print("Root Mean Squared Error : " + (metrics.mean_squared_error(y_test, y_test_predict)**0.5).astype(str) )

# Ridge Regression

rdg = linear_model.Ridge(alpha=0.01)
rdg.fit(X_train, y_train)

y_test_predict = rdg.predict(X_test)
y_train_predict = rdg.predict(X_train)

print('R^2 training: %.3f, R^2 test: %.3f' % (
      (metrics.r2_score(y_train, y_train_predict)), 
      (metrics.r2_score(y_test, y_test_predict))))

print("Root Mean Squared Error : " + (metrics.mean_squared_error(y_test, y_test_predict)**0.5).astype(str) )

# Random Forest

rf = RandomForestRegressor(n_estimators=100, max_features=3, max_depth=8, n_jobs=-1, random_state=0) 

rf.fit(X_train, y_train)

y_test_predict = rf.predict(X_test)
y_train_predict = rf.predict(X_train)

print('R^2 training: %.3f, R^2 test: %.3f' % (
      (metrics.r2_score(y_train, y_train_predict)), 
      (metrics.r2_score(y_test, y_test_predict))))

print("Root Mean Squared Error : " + (metrics.mean_squared_error(y_test, y_test_predict)**0.5).astype(str) )

# Output a pickle file for the model
# joblib.dump(rf, 'FD00X.pkl') 

# Decision Tree

dtrg = DecisionTreeRegressor(max_depth=9, random_state=123)
dtrg.fit(X_train, y_train)

y_test_predict = dtrg.predict(X_test)
y_train_predict = dtrg.predict(X_train)

print('R^2 training: %.3f, R^2 test: %.3f' % (
      (metrics.r2_score(y_train, y_train_predict)), 
      (metrics.r2_score(y_test, y_test_predict))))

print("Root Mean Squared Error : " + (metrics.mean_squared_error(y_test, y_test_predict)**0.5).astype(str) )

# Test Prediction
model = joblib.load(open('FD003.pkl','rb'))
print(model.predict([[25.001,0.6215,60,462.54,536.71,1262.32,1052.92,7.05,9.03,175.72,1915.33,8010.74,0.94,36.74,164.72,2028.17,7876.61,10.8998,0.02,308,1915,84.93,14.32,8.5193
]]))
