==> 3 - Deploy Web Service.ipynb <==
# import required azure library
from azureml.core.model import Model
from azureml.core import Workspace
from azureml.core.resource_configuration import ResourceConfiguration
from azureml.core import Webservice
from azureml.exceptions import WebserviceException
from azureml.core.webservice import AciWebservice, Webservice

# load workspace configuration from the config.json file in the current folder.
ws = Workspace.from_config()
print(ws.name, ws.resource_group, ws.location, ws.subscription_id, sep='\n')

# Register the model into Azure Machine Learning
model = Model.register(model_path = "./FD001.pkl",
                       model_name = "Rf",
                       model_framework=Model.Framework.SCIKITLEARN,
                       description = "SmartMaintenance ML Model",
                       workspace = ws)

deployment_config = AciWebservice.deploy_configuration(cpu_cores = 1, memory_gb = 1)
service = Model.deploy(ws, "aciservice", [model], deployment_config)
service.wait_for_deployment(show_output = True)
print(service.state)

import sklearn

print('The scikit-learn version is {}.'.format(sklearn.__version__))



# Azure Kubernetes Service (AKS)	
Real-time inference	Yes (web service deployment)		
- Use for high-scale production deployments. Provides fast response time and autoscaling of the deployed service. Cluster autoscaling isn't supported through the Azure Machine Learning SDK. To change the nodes in the AKS cluster, use the UI for your AKS cluster in the Azure portal. AKS is the only option available for the designer.

# Azure Container (ACI)
Instances	Testing or development	 	 	
- Use for low-scale CPU-based workloads that require less than 48 GB of RAM.

<br/>

Choose
Compute Type
- ACI
- Disable authentication

Wait for REST endpoint <br/>
Send a POST request in Postman to test it
